requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
decorator==4.4.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.53.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
imageio==2.26.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.24.2
oauthlib==3.2.2
Pillow==9.4.0
proglog==0.1.10
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
urllib3==1.26.15
